remove Go 1.12 code
